Test with alkiln v3 action-controlled workflow only
This uses only the workflow file to trigger ALKiln tests. Removed package.json and other npm files and files for local development as the ALKiln will manage these now. References the ALKiln v3 branch to keep up with updates to that branch.

Also fixes the file named in the test to match the current main file name.